Changed to maintain module.exports structure similar to the main project
